rename container of directions for an intersection into
IntersectionDirections instead of ConflictingDirections, as it contains
not only conflicting but all directions per intersection
